Move favorites into separate crate
