+++ nym-vpn-core/crates/nym-favorites/src/util.rs
