+++ nym-vpn-core/crates/nym-vpn-lib/src/lib.rs
@@ -9,7 +9,6 @@ mod bandwidth_monitor;
 pub mod cache_refresh;
 pub mod config;
 mod dns_filter;
-mod favorites;
 pub mod logging;
 mod mixnet;
 pub mod privy;
@@ -43,7 +42,6 @@ pub use nym_sdk::{
 };
 
 pub use crate::{
-    favorites::{FavoritesManager, RecentGatewayCache, RecentsError, RecentsManager},
     mixnet::{
         DEFAULT_MIN_GATEWAY_PERFORMANCE, DEFAULT_MIN_MIXNODE_PERFORMANCE, MixnetError,
         VpnTopologyProvider, VpnTopologyService, VpnTopologyServiceError, VpnTopologyServiceHandle,
